requirements.txt
sentence-transformers
google-generativeai
scikit-learn
pyngrok
flask_cors